# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rehearsed",
    "description": "A production-ready FastAPI template with LangGraph and Langfuse integration",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/chatbot/chat": {
      "post": {
        "tags": [
          "chatbot"
        ],
        "summary": "Chat",
        "description": "Process a chat request using LangGraph.\n\nArgs:\n    request: The FastAPI request object for rate limiting.\n    chat_request: The chat request containing messages.\n    background_tasks: The background tasks to run.\n    session: The current session from the auth token.\n    text_to_speech_service: The text-to-speech service instance.\n\nReturns:\n    ChatResponse: The processed chat response.\n\nRaises:\n    HTTPException: If there's an error processing the request.",
        "operationId": "chat_api_v1_chatbot_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/chatbot/chat/stream": {
      "post": {
        "tags": [
          "chatbot"
        ],
        "summary": "Chat Stream",
        "description": "Process a chat request using LangGraph with streaming response.\n\nArgs:\n    request: The FastAPI request object for rate limiting.\n    chat_request: The chat request containing messages.\n    session: The current session from the auth token.\n    text_to_speech_service: The text-to-speech service instance.\n\nReturns:\n    StreamingResponse: A streaming response of the chat completion.\n\nRaises:\n    HTTPException: If there's an error processing the request.",
        "operationId": "chat_stream_api_v1_chatbot_chat_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/chatbot/messages": {
      "get": {
        "tags": [
          "chatbot"
        ],
        "summary": "Get Session Messages",
        "description": "Get all messages for a session.\n\nArgs:\n    request: The FastAPI request object for rate limiting.\n    session: The current session from the auth token.\n    text_to_speech_service: The text-to-speech service instance.\n\nReturns:\n    ChatResponse: All messages in the session.\n\nRaises:\n    HTTPException: If there's an error retrieving the messages.",
        "operationId": "get_session_messages_api_v1_chatbot_messages_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "chatbot"
        ],
        "summary": "Clear Chat History",
        "description": "Clear all messages for a session.\n\nArgs:\n    request: The FastAPI request object for rate limiting.\n    session: The current session from the auth token.\n\nReturns:\n    dict: A message indicating the chat history was cleared.",
        "operationId": "clear_chat_history_api_v1_chatbot_messages_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/chatbot/feedback/{feedback_id}": {
      "get": {
        "tags": [
          "chatbot"
        ],
        "summary": "Get Feedback",
        "description": "Get async inline feedback by ID.\n\nThis endpoint is used for polling feedback that is generated asynchronously\nafter the chat response is returned.\n\nArgs:\n    request: The FastAPI request object for rate limiting.\n    feedback_id: The unique ID of the feedback request.\n    session: The current session from the auth token.\n\nReturns:\n    dict: The feedback status and content if ready.\n        - status: \"pending\" | \"ready\" | \"failed\"\n        - feedback: List of feedback strings (empty if not ready)",
        "operationId": "get_feedback_api_v1_chatbot_feedback__feedback_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "feedback_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Feedback Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Agent": {
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "default": ""
          },
          "scenario_id": {
            "type": "integer",
            "title": "Scenario Id"
          },
          "voice_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice Id"
          },
          "display_text_color": {
            "type": "string",
            "title": "Display Text Color",
            "default": ""
          },
          "avatar_gcs_uri": {
            "type": "string",
            "title": "Avatar Gcs Uri",
            "default": ""
          },
          "objective": {
            "type": "string",
            "title": "Objective",
            "default": ""
          },
          "instructions": {
            "type": "string",
            "title": "Instructions",
            "default": ""
          },
          "constraints": {
            "type": "string",
            "title": "Constraints",
            "default": ""
          },
          "context": {
            "type": "string",
            "title": "Context",
            "default": ""
          },
          "agent_personality_id": {
            "type": "integer",
            "title": "Agent Personality Id"
          },
          "owner_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "scenario_id",
          "agent_personality_id"
        ],
        "title": "Agent",
        "description": "Agent model for storing chat agent prompts and metadata.\n\nAttributes:\n    id: The primary key\n    name: Name of the session (defaults to empty string)\n    created_at: When the session was created\n    messages: Relationship to session messages\n    user: Relationship to the session owner"
      },
...
      "AgentPersonality": {
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "personality_description": {
            "type": "string",
            "title": "Personality Description"
          },
          "owner_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "personality_description"
        ],
        "title": "AgentPersonality",
        "description": "Database model for agent personality configurations."
      },
...
      "ChatRequest": {
        "properties": {
          "messages": {
            "items": {
              "$ref": "#/components/schemas/Message"
            },
            "type": "array",
            "title": "Messages",
            "description": "List of messages in the conversation"
          },
          "is_resumption": {
            "type": "boolean",
            "title": "Is Resumption",
            "description": "Whether the conversation is being resumed",
            "default": false
          },
          "resumption_text": {
            "type": "string",
            "title": "Resumption Text",
            "description": "The text to resume the conversation",
            "default": ""
          },
          "audio_base64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Base64",
            "description": "Optional base64-encoded audio bytes for speech-to-text transcription"
          }
        },
        "type": "object",
        "required": [
          "messages"
        ],
        "title": "ChatRequest",
        "description": "Request model for chat endpoint.\n\nAttributes:\n    messages: List of messages in the conversation.\n    is_resumption: Whether the conversation is being resumed.\n    resumption_text: The text to resume the conversation.\n    audio_base64: Optional base64-encoded audio bytes for speech-to-text transcription."
      },
      "ChatResponse": {
        "properties": {
          "messages": {
            "items": {
              "$ref": "#/components/schemas/Message"
            },
            "type": "array",
            "title": "Messages",
            "description": "List of messages in the conversation"
          },
          "interrupt_task": {
            "type": "string",
            "title": "Interrupt Task",
            "description": "The task to interrupt the conversation",
            "default": ""
          },
          "interrupt_value": {
            "type": "string",
            "title": "Interrupt Value",
            "description": "The value to interrupt the conversation",
            "default": ""
          },
          "interrupt_value_type": {
            "type": "string",
            "enum": [
              "text",
              "image",
              "audio",
              "video"
            ],
            "title": "Interrupt Value Type",
            "description": "The type of the interrupt value",
            "default": "text"
          },
          "student_responses": {
            "items": {
              "$ref": "#/components/schemas/StudentResponse"
            },
            "type": "array",
            "title": "Student Responses",
            "description": "List of student responses"
          },
          "inline_feedback": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Inline Feedback",
            "description": "List of inline feedback"
          },
          "feedback_request_id": {
            "type": "string",
            "title": "Feedback Request Id",
            "description": "ID to poll for async inline feedback if not immediately available",
            "default": ""
          },
          "summary_feedback": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SummaryFeedbackResponse"
              },
              {
                "type": "string"
              }
            ],
            "title": "Summary Feedback",
            "description": "The summary feedback for the student responses",
            "default": ""
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "description": "The summary of the student responses",
            "default": ""
          },
          "answering_student": {
            "type": "integer",
            "title": "Answering Student",
            "description": "The student number that is answering",
            "default": 0
          },
          "appropriate_response": {
            "type": "boolean",
            "title": "Appropriate Response",
            "description": "Whether the response is appropriate",
            "default": false
          },
          "appropriate_explanation": {
            "type": "string",
            "title": "Appropriate Explanation",
            "description": "The explanation for why the response is appropriate",
            "default": ""
          },
          "learning_goals_achieved": {
            "type": "boolean",
            "title": "Learning Goals Achieved",
            "description": "Whether the learning goals were achieved",
            "default": false
          },
          "transcribed_text": {
            "type": "string",
            "title": "Transcribed Text",
            "description": "The transcribed text from audio input",
            "default": ""
          },
          "interrupt": {
            "items": {
              "$ref": "#/components/schemas/Interrupt"
            },
            "type": "array",
            "title": "Interrupt",
            "description": "List of interrupts"
          }
        },
        "type": "object",
        "title": "ChatResponse",
        "description": "Response model for chat endpoint.\n\nAttributes:\n    messages: List of messages in the conversation."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Interrupt": {
        "properties": {
          "value": {
            "title": "Value"
          },
          "id": {
            "type": "string",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "value",
          "id"
        ],
        "title": "Interrupt"
      },
...
      "Message": {
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "user",
              "assistant",
              "system"
            ],
            "title": "Role",
            "description": "The role of the message sender"
          },
          "content": {
            "type": "string",
            "maxLength": 3000,
            "minLength": 1,
            "title": "Content",
            "description": "The content of the message"
          }
        },
        "type": "object",
        "required": [
          "role",
          "content"
        ],
        "title": "Message",
        "description": "Message model for chat endpoint.\n\nAttributes:\n    role: The role of the message sender (user or assistant).\n    content: The content of the message."
      },
...
      "StudentResponse": {
        "properties": {
          "student_response": {
            "type": "string",
            "title": "Student Response",
            "description": "The response from the student"
          },
          "student_details": {
            "$ref": "#/components/schemas/Agent"
          },
          "student_personality": {
            "$ref": "#/components/schemas/AgentPersonality"
          },
          "audio_base64": {
            "type": "string",
            "title": "Audio Base64",
            "description": "Base64-encoded audio bytes for the student response",
            "default": ""
          },
          "audio_id": {
            "type": "string",
            "title": "Audio Id",
            "description": "Identifier for fetching audio lazily",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "student_response",
          "student_details",
          "student_personality"
        ],
        "title": "StudentResponse",
        "description": "Model for the student response."
      },
...
      "SummaryFeedbackResponse": {
        "properties": {
          "lesson_summary": {
            "type": "string",
            "title": "Lesson Summary",
            "description": "What happened in the lesson"
          },
          "key_moments": {
            "type": "string",
            "title": "Key Moments",
            "description": "Two to three things they did excellent in the lesson"
          },
          "overall_feedback": {
            "type": "string",
            "title": "Overall Feedback",
            "description": "Briefly describe how they did overall in the lesson"
          },
          "your_strengths": {
            "type": "string",
            "title": "Your Strengths",
            "description": "What they did well in the lesson"
          },
          "areas_for_growth": {
            "type": "string",
            "title": "Areas For Growth",
            "description": "What they could improve in the lesson"
          },
          "next_steps": {
            "type": "string",
            "title": "Next Steps",
            "description": "What they should try next in the lesson to demonstrate the skill in the future"
          },
          "celebration": {
            "type": "string",
            "title": "Celebration",
            "description": "What they should celebrate in the lesson"
          }
        },
        "type": "object",
        "required": [
          "lesson_summary",
          "key_moments",
          "overall_feedback",
          "your_strengths",
          "areas_for_growth",
          "next_steps",
          "celebration"
        ],
        "title": "SummaryFeedbackResponse",
        "description": "Model for the summary feedback response."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}